Advances in IBM quantum processor
On January 1st, 2017 IBM unveiled QASM v2 and further advances in the physical capabilities of the processor.  The ability for two additional qbits to become entangled provides greater capabilities in algorithmic design and implementations.  I will be updating the quantum search algorithm in this repo to reflect these new advances from IBM.